Start of producing scaled back stems per slide
