+++ .gitignore
@@ -29,3 +29,5 @@ yarn-error.log*
 
 # local env files
 .env
+
+.claude
